modular routing for /notes
